FFT-backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
qrcode==8.1
Pillow==11.2.1
python-multipart==0.0.20
anthropic==0.52.0
pyzbar==0.1.9
psycopg[binary]==3.2.6
python-dotenv==1.0.1
